Handling expressions like foo[bar].baz
This change makes it possible to use expressions like `foo[bar].baz`. In
order to make this work, the parser was updated to allow optional spaces
after `]` and then splitting the bracket and any spaces into separate
tokens in the `tokenize` function. This is needed in order do
differentiate `foo[bar].baz` from `foo[bar] baz` as in `eq foo[bar]
baz`. It is not possible to do this using regular expressions alone
because you cannot match a space only when it follows a `]` (regex in
JavaScript do not support look-behind).